requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
dj-database-url>=2.1.0
gunicorn>=21.2.0
# PostgreSQL adapter (psycopg3 - modern, supports Python 3.13+)
psycopg[binary]>=3.2.0
# Redis support (for Phase 2)
redis>=5.0.0
django-redis>=5.4.0
# Django Channels for WebSockets (Phase 2)
channels>=4.0.0
channels-redis>=4.2.0
# Celery for async tasks (Phase 2)
celery>=5.3.0
# AI/ML libraries (Phase 2)
openai>=1.0.0
langchain>=0.1.0
# Image processing (for posts)
Pillow>=10.0.0
# Utilities
numpy>=1.24.0
scikit-learn>=1.3.0

